--- a/src/crs/basic/cartesian.jl
+++ b/src/crs/basic/cartesian.jl
@@ -121,16 +121,6 @@ lentype(::Type{<:Cartesian{Datum,N,L}}) where {Datum,N,L} = L
 
 ==(coords₁::Cartesian{Datum,N}, coords₂::Cartesian{Datum,N}) where {Datum,N} = _coords(coords₁) == _coords(coords₂)
 
-function Base.isapprox(coords₁::C, coords₂::C; kwargs...) where {C<:Cartesian}
-  tuple₁ = _coords(coords₁)
-  tuple₂ = _coords(coords₂)
-  all(1:length(tuple₁)) do i
-    c₁ = getfield(tuple₁, i)
-    c₂ = getfield(tuple₂, i)
-    isapprox(c₁, c₂; rtol=rtol(c₁), atol=atol(c₁), kwargs...)
-  end
-end
-
 Random.rand(rng::Random.AbstractRNG, ::Type{Cartesian{Datum,N}}) where {Datum,N} =
   Cartesian{Datum}(ntuple(i -> rand(rng), N)...)